backend/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0

# Database - MySQL
sqlalchemy==2.0.25
aiomysql==0.2.0
cryptography>=41.0.0
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Task Queue (Optional - comment out if not needed immediately)
celery==5.3.6

# Basic Data Processing
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0

# Monitoring & Logging
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
pytz==2024.1

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4

# Development
black==24.1.1

# ===== OPTIONAL: ML/AI Libraries (Install separately if needed) =====
# Uncomment and install individually based on your needs:
# 
# Machine Learning:
# scikit-learn>=1.3.0
# xgboost>=2.0.0
# lightgbm>=4.0.0
#
# Deep Learning (large downloads):
# tensorflow>=2.13.0  # or tensorflow-cpu for CPU-only
# torch>=2.0.0  # PyTorch
#
# NLP:
# transformers>=4.30.0
# sentence-transformers>=2.0.0
# spacy>=3.6.0
#
# AWS:
# boto3>=1.34.0
# aioboto3>=12.0.0
